+++ crates/testing/tests/tests_1/consensus_task.rs
@@ -14,10 +14,9 @@ use hotshot_testing::{
         exact, quorum_proposal_send, quorum_proposal_validated, quorum_vote_send, timeout_vote_send,
     },
     script::{run_test_script, TestScriptStage},
-    task_helpers::{
-        build_system_handle, vid_share, key_pair_for_id, vid_scheme_from_view_number,
+    helpers::{
+        build_system_handle, vid_share, key_pair_for_id, vid_scheme_from_view_number, permute_input_with_index_order
     },
-    test_helpers::permute_input_with_index_order,
     view_generator::TestViewGenerator,
 };
 use hotshot_types::{
@@ -123,7 +122,7 @@ async fn test_consensus_vote() {
     use hotshot_task_impls::{consensus::ConsensusTaskState, events::HotShotEvent::*};
     use hotshot_testing::{
         script::{run_test_script, TestScriptStage},
-        task_helpers::build_system_handle,
+        helpers::build_system_handle,
         view_generator::TestViewGenerator,
     };
