+++ crates/testing/tests/tests_1/proposal_ordering.rs
@@ -11,8 +11,7 @@ use hotshot_example_types::{
 use hotshot_task_impls::{consensus::ConsensusTaskState, events::HotShotEvent::*};
 use hotshot_testing::{
     predicates::event::{all_predicates, exact, quorum_proposal_send, quorum_proposal_validated},
-    task_helpers::{vid_share, vid_scheme_from_view_number},
-    test_helpers::permute_input_with_index_order,
+    helpers::{vid_share, vid_scheme_from_view_number, permute_input_with_index_order},
     view_generator::TestViewGenerator,
 };
 use hotshot_types::{
@@ -29,7 +28,7 @@ use sha2::Digest;
 async fn test_ordering_with_specific_order(input_permutation: Vec<usize>) {
     use hotshot_testing::{
         script::{run_test_script, TestScriptStage},
-        task_helpers::build_system_handle,
+        helpers::build_system_handle,
     };
 
     async_compatibility_layer::logging::setup_logging();
