+++ crates/testing/tests/tests_1/view_sync_task.rs
@@ -10,7 +10,7 @@ use hotshot_types::{data::ViewNumber, traits::node_implementation::ConsensusTime
 #[cfg_attr(async_executor_impl = "async-std", async_std::test)]
 async fn test_view_sync_task() {
     use hotshot_task_impls::{harness::run_harness, view_sync::ViewSyncTaskState};
-    use hotshot_testing::task_helpers::build_system_handle;
+    use hotshot_testing::helpers::build_system_handle;
     use hotshot_types::simple_vote::ViewSyncPreCommitData;
 
     async_compatibility_layer::logging::setup_logging();
